Dirty fix for overlapping paths
Sometimes a file (with no extension) is seen first then another file that has an overlapping path is seen e.g.
File: A/B creates file A/B
File: A/B/C.ext fails because A/B is not a directory

Sometimes the file with no extension is seen second e.g.
File: A/B/C.ext creates directory A/B and file A/B/C.ext
File: A/B fails because A/B is a directory and this tried to create a file A/B

This fix simply ignores the second file in both cases. A more robust fix would be to detect the overlapping extensionless file and give it a custom extenstion e.g. A/B.null.